Fix network dependency injection
