+++ app/src/main/java/com/zenika/zenilunch/data/network/RestaurantGitHubApi.kt
@@ -1,0 +1,9 @@
+package com.zenika.zenilunch.data.network
+
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface RestaurantGitHubApi {
+    @GET("{file}")
+    suspend fun getRestaurants(@Path("file") file: String): List<RestaurantDto>
+}
