+++ app/src/main/java/com/zenika/zenilunch/suggestion/SuggestionViewModel.kt
@@ -6,7 +6,7 @@ import com.zenika.zenilunch.RestaurantUIModel
 import com.zenika.zenilunch.domain.GetSuggestionsUseCase
 import com.zenika.zenilunch.domain.NB_SUGGESTIONS_DESIRED
 import com.zenika.zenilunch.mapper.convertRestaurantObject
-import com.zenika.zenilunch.network.RestaurantDto
+import com.zenika.zenilunch.data.network.RestaurantDto
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.toImmutableList
