+++ app/src/test/java/com/zenika/zenilunch/GetRestaurantsSortedByNameUseCaseTest.kt
@@ -3,19 +3,21 @@ package com.zenika.zenilunch
 import com.zenika.zenilunch.agency.model.Agency
 import com.zenika.zenilunch.agency.model.LatLng
 import com.zenika.zenilunch.domain.GetRestaurantsSortedByNameUseCase
-import com.zenika.zenilunch.network.RestaurantDto
+import com.zenika.zenilunch.data.network.RestaurantDto
 import com.zenika.zenilunch.repository.AgencyRepository
 import com.zenika.zenilunch.repository.RestaurantRepository
 import io.mockk.coEvery
 import io.mockk.impl.annotations.InjectMockKs
 import io.mockk.impl.annotations.MockK
 import io.mockk.junit5.MockKExtension
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 
+@OptIn(ExperimentalCoroutinesApi::class)
 @ExtendWith(MockKExtension::class)
 internal class GetRestaurantsSortedByNameUseCaseTest {
