+++ app/src/test/java/com/zenika/zenilunch/repository/RestaurantRepositoryTest.kt
@@ -4,12 +4,13 @@ import com.zenika.zenilunch.agency.model.Agency
 import com.zenika.zenilunch.agency.model.LatLng
 import com.zenika.zenilunch.data.HiddenRestaurant
 import com.zenika.zenilunch.data.RestaurantDao
-import com.zenika.zenilunch.network.RestaurantDto
-import com.zenika.zenilunch.network.RestaurantNetwork
+import com.zenika.zenilunch.data.network.RestaurantDto
+import com.zenika.zenilunch.data.network.RestaurantGitHubApi
 import io.mockk.coEvery
 import io.mockk.impl.annotations.InjectMockKs
 import io.mockk.impl.annotations.MockK
 import io.mockk.junit5.MockKExtension
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.DisplayName
@@ -17,6 +18,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import java.util.Date
 
+@OptIn(ExperimentalCoroutinesApi::class)
 @ExtendWith(MockKExtension::class)
 class RestaurantRepositoryTest {
 
@@ -24,7 +26,7 @@ class RestaurantRepositoryTest {
     lateinit var restaurantDao: RestaurantDao
 
     @MockK
-    lateinit var restaurantNetwork: RestaurantNetwork
+    lateinit var restaurantGitHubApi: RestaurantGitHubApi
 
     @InjectMockKs
     lateinit var repository: RestaurantRepository
@@ -34,7 +36,7 @@ class RestaurantRepositoryTest {
     fun shouldNotReturnAnyHiddenRestaurant() = runTest {
         coEvery { restaurantDao.getHiddenRestaurants() } returns
             listOf(HiddenRestaurant("Pas bon", Date()))
-        coEvery { restaurantNetwork.getRestaurants(any()) } returns
+        coEvery { restaurantGitHubApi.getRestaurants(any()) } returns
             listOf(
                 RestaurantDto("Pas bon", "Pâtes", "##€", false, false, 0.0, 0.0),
                 RestaurantDto("Super bon", "Pizza", "€", true, true, 0.0, 0.0),
